Tier 0 hardening consolidation:
- Durable artifact /tmp deny gate (moved before training — B5 fix)
- RunPod lifecycle helper + per-request executionTimeout policy (B4 fix)
- Slim training image (Dockerfile.slim, drops torch)
- Metric sanity bounds (flag implausible Sharpe/returns/drawdown)
- Ruff burn-down: 1845 to 249 errors (86.5% reduction)
- 139 tests pass (7+4+23+46+18+41)

Receipts: reports/tier0-consolidation/

+++ services/api/src/api/openbb_health_store.py
@@ -158,7 +158,7 @@ def summarise(entries: list[dict[str, Any]]) -> dict[str, Any]:
     p95: int | None = None
     if latencies:
         p50 = latencies[len(latencies) // 2]
-        p95_index = max(0, int(round(len(latencies) * 0.95)) - 1)
+        p95_index = max(0, round(len(latencies) * 0.95) - 1)
         p95 = latencies[min(p95_index, len(latencies) - 1)]
     last_error = next(
         (e["error_type"] for e in reversed(entries) if not e["ok"]),
